% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder.R
\name{create_or_rewrite_output_folder}
\alias{create_or_rewrite_output_folder}
\title{Create or Rewrite Output Folder}
\usage{
create_or_rewrite_output_folder(folder_path = "output")
}
\arguments{
\item{folder_path}{The path of the folder to create or clear. Default is "output".}
}
\value{
The absolute path of the created or cleared folder.
}
\description{
This function creates a new folder at the specified path or clears an existing folder
if it already exists.
}
\examples{
create_or_rewrite_output_folder("path/to/output")

}
